requirements.txt
tensorflow<2.11
pandas
numpy<2.0.0
opencv-python
matplotlib
seaborn
scikit-learn
pillow
tqdm
